fix(cli): absorb #2017 — webhooks allowlist, readme.md rule exclusion, move isFactFile test
- migrate.ts SHIPPED_ALLOWLIST: add webhooks (canonical shipped resource, fixes #2017 false-positive orphan warning)
- memory.ts RULE_FILE_NAMES: add readme.md so directory README files are treated as metadata, not user facts
- memory.test.ts: move isFactFile regression test here from humans.test.ts (1:1 source/test rule); covers AGENTS.md, README.md, CLAUDE.md, GEMINI.md, MEMORY.md — asserts only my-fact.md is listed
- humans.test.ts: remove misplaced isFactFile test; no imports became unused
