+++ apps/cli/src/lib/memory.ts
@@ -78,7 +78,7 @@ export function ensureUserMemoryDir(): string {
   return dir;
 }
 
-const RULE_FILE_NAMES = new Set(['memory.md', 'agents.md', 'claude.md', 'gemini.md']);
+const RULE_FILE_NAMES = new Set(['memory.md', 'agents.md', 'claude.md', 'gemini.md', 'readme.md']);
 
 function isFactFile(name: string): boolean {
   return name.endsWith('.md') && !RULE_FILE_NAMES.has(name.toLowerCase());
